refactored-chatcontroller and chatservice
